Show type on cursor hover
This brings some Visual Studio vibes, when the cursor is on a type of
function, shows its type. CocActionAsync('doHover') is binded on 'w'
(what), and shows whenever used so this might be a little redundant and
also noisy. I'll keep it around because it's really cool. Does it work
with mouse independent from cursor?